requirements.txt
Django==2.1.2
#DRF
djangorestframework